fix error handling of call to certopensystemstorea
CertOpenSystemStoreA returns null when it fails. This can e.g. be due to
insufficient access rights to the certificate store. However, ffi
returns a buffer whose truthiness will always be true. To properly catch
the error and prevent the application from crashing, we need to check for
null using the ref package.